requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
Flask==3.1.2
frozenlist==1.8.0
genai==2.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
multidict==6.7.0
numpy==2.2.6
openai==0.27.10
pandas==2.3.3
parso==0.8.5
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
playwright==1.56.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.5
pypdfium2==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
yarl==1.22.0
